// src/gql/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Abilities = {
  __typename?: 'Abilities';
  /** The first ability of a Pokémon */
  first: Ability;
  /** The hidden ability of a Pokémon */
  hidden?: Maybe<Ability>;
  /** The second ability of a Pokémon */
  second?: Maybe<Ability>;
  /** The special ability of a Pokémon */
  special?: Maybe<Ability>;
};
// ...
export type Ability = {
  __typename?: 'Ability';
  /** Bulbapedia page for an ability */
  bulbapediaPage: Scalars['String']['output'];
  /** The long description for an ability */
  desc?: Maybe<Scalars['String']['output']>;
  /** Whether this ability has effects outside of battle, and if so what the effect is */
  isFieldAbility?: Maybe<Scalars['String']['output']>;
  /** The key of the ability as stored in the API */
  key: AbilitiesEnum;
  /** The name for an ability */
  name: Scalars['String']['output'];
  /**
   * The Pokémon that have this ability.
   * Note that when querying this field and nesting deep into Pokemon.abilities, that the nested list
   * will not have any values to prevent infinite looping data.
   */
  pokemonThatHaveThisAbility: Array<Maybe<Pokemon>>;
  /** Serebii page for an ability */
  serebiiPage: Scalars['String']['output'];
  /** The short description for an ability */
  shortDesc: Scalars['String']['output'];
  /** Smogon page for an ability */
  smogonPage: Scalars['String']['output'];
};

/** A Pokémon catch rate entry */
export type CatchRate = {
  __typename?: 'CatchRate';
  /** The base catch rate for a Pokémon that will be used to calculate the final catch rate */
  base: Scalars['Int']['output'];
  /** The chance to capture a Pokémon when an ordinary Poké Ball is thrown at full health without any status condition */
  percentageWithOrdinaryPokeballAtFullHealth: Scalars['String']['output'];
};
// ...
export type Flavor = {
  __typename?: 'Flavor';
  /** The flavor text for this entry */
  flavor: Scalars['String']['output'];
  /** The name of the game this flavor text is from */
  game: Scalars['String']['output'];
};
// ...
export type Pokemon = {
  __typename?: 'Pokemon';
  /** The abilities for a Pokémon */
  abilities: Abilities;
  /** The back sprite for a Pokémon. For most Pokémon this will be the animated gif, with some exceptions that were older-gen exclusive */
  backSprite: Scalars['String']['output'];
  /** Base form if this entry describes an alternate form */
  baseForme?: Maybe<Scalars['String']['output']>;
  /** Base species if this entry describes a special form */
  baseSpecies?: Maybe<Scalars['String']['output']>;
  /** Base stats for a Pokémon */
  baseStats: Stats;
  /** The total of all base stats for a Pokémon */
  baseStatsTotal: Scalars['Int']['output'];
  /** Bulbapedia page for a Pokémon */
  bulbapediaPage: Scalars['String']['output'];
  /** The catch rate data for a Pokémon */
  catchRate?: Maybe<CatchRate>;
  /** The colour of a Pokémon as listed in the Pokedex */
  color: Scalars['String']['output'];
  /** Any other *cosmetic* forms for a Pokémon, distinguished from other formes as cosmetic formes only change the look of the Pokémon, while other formes might also change an ability, move set or other data. */
  cosmeticFormes?: Maybe<Array<Scalars['String']['output']>>;
  /** The egg groups a Pokémon is in */
  eggGroups?: Maybe<Array<Scalars['String']['output']>>;
  /** EV yields for a Pokémon */
  evYields: EvYields;
  /** The evolution level, or special method, for a Pokémon */
  evolutionLevel?: Maybe<Scalars['String']['output']>;
  /** The evolutions for a Pokémon, if any */
  evolutions?: Maybe<Array<Pokemon>>;
  /** The flavor texts for a Pokémon */
  flavorTexts: Array<Flavor>;
  /** The form identifier of a Pokémon */
  forme?: Maybe<Scalars['String']['output']>;
  /** The single letter identifier of the form */
  formeLetter?: Maybe<Scalars['String']['output']>;
  /** The gender data for a Pokémon */
  gender: Gender;
  /** The height of a Pokémon in meters */
  height: Scalars['Float']['output'];
  /** Whether the egg of a Pokémon is obtainable */
  isEggObtainable: Scalars['Boolean']['output'];
  /** The key of the Pokémon as stored in the API */
  key: PokemonEnum;
  /** The learnset for this pokemon */
  learnsets?: Maybe<GenerationalPokemonLearnset>;
  /** Whether this Pokémon is a legendary Pokémon. The list is based on what is provided by Bulbapedia. */
  legendary: Scalars['Boolean']['output'];
  /** The levelling rate of a Pokémon */
  levellingRate?: Maybe<Scalars['String']['output']>;
  /** The maximum number of steps required for the egg of a Pokémon to hatch */
  maximumHatchTime?: Maybe<Scalars['Int']['output']>;
  /** The minimum number of steps required for the egg of a Pokémon to hatch */
  minimumHatchTime?: Maybe<Scalars['Int']['output']>;
  /** Whether this Pokémon is a mythical Pokémon. The list is based on what is provided by Bulbapedia. */
  mythical: Scalars['Boolean']['output'];
  /** The dex number for a Pokémon */
  num: Scalars['Int']['output'];
  /** Any other forms for a Pokémon */
  otherFormes?: Maybe<Array<Scalars['String']['output']>>;
  /** The preevolutions for a Pokémon, if any */
  preevolutions?: Maybe<Array<Pokemon>>;
  /** Serebii page for a Pokémon */
  serebiiPage: Scalars['String']['output'];
  /** The shiny back sprite for a Pokémon. For most Pokémon this will be the animated gif, with some exceptions that were older-gen exclusive */
  shinyBackSprite: Scalars['String']['output'];
  /** The shiny sprite for a Pokémon. For most Pokémon this will be the animated gif, with some exceptions that were older-gen exclusive */
  shinySprite: Scalars['String']['output'];
  /** Smogon page for a Pokémon */
  smogonPage: Scalars['String']['output'];
  /** The smogon tier a Pokémon falls under */
  smogonTier: Scalars['String']['output'];
  /** The species name for a Pokémon */
  species: Scalars['String']['output'];
  /** The sprite for a Pokémon. For most Pokémon this will be the animated gif, with some exceptions that were older-gen exclusive */
  sprite: Scalars['String']['output'];
  /** The types for a Pokémon */
  types: Array<PokemonType>;
  /** The weight of a Pokémon in kilograms */
  weight: Scalars['Float']['output'];
};
// ...
export type PokemonType = {
  __typename?: 'PokemonType';
  /** The type matchup for this type */
  matchup: TypeMatchup;
  /** The name of the typ */
  name: Scalars['String']['output'];
};

export type Query = {
  __typename?: 'Query';
  /** Gets the details on a Pokémon ability, using the ability name */
  getAbility: Ability;
  /**
   * Returns a list of all the known Pokémon.
   *
   * For every Pokémon all the data on each requested field is returned.
   *
   * **_NOTE:_ To skip all CAP Pokémon, PokéStar Pokémon, Missingno, and 'M (00) provide an `offset` of 89**
   *
   * You can provide `take` to limit the amount of Pokémon to return (default: 1), set the offset of where to start with `offset`, and reverse the entire array with `reverse`.
   *
   * You can provide `takeFlavorTexts` to limit the amount of flavour texts to return, set the offset of where to start with `offsetFlavorTexts`, and reverse the entire array with `reverseFlavorTexts`.
   *
   * While the API will currently not rate limit the usage of this query, it may do so in the future.
   *
   * It is advisable to cache responses of this query.
   */
  getAllPokemon: Array<Pokemon>;
  /**
   * Gets details on a Pokémon ability, using a fuzzy search on name
   *
   * This can be used to find multiple results based on the query
   *
   * You can provide `take` to limit the amount of abilities to return (default: 1), set the offset of where to start with `offset`, and reverse the entire array with `reverse`.
   */
  getFuzzyAbility: Array<Ability>;
  /**
   * Gets details on a Pokémon item, using a fuzzy search on name
   *
   * This can be used to find multiple results based on the query
   *
   * By default only 1 result is returned. You can provide the arguments `take`, `offset`, and `reverse` to modify this behaviour.
   */
  getFuzzyItem: Array<Item>;
  /**
   * Gets the learnset for a given Pokémon and move.
   *
   * A fuzzy search is performed to find a matching Pokémon and move
   *
   * Multiple moves are possible by putting them in an array: `[move1, move2]`.
   *
   * You can also apply a generation filter (only results for the given generation will be returned) with the generation argument
   */
  getFuzzyLearnset: Learnset;
  /**
   * Gets details on a Pokémon move, using a fuzzy search on name
   *
   * This can be used to find multiple results based on the query
   *
   * By default only 1 result is returned. You can provide the arguments `take`, "offset", and "reverse" to modify this behaviour.
   */
  getFuzzyMove: Array<Move>;
  /**
   * Gets details on one or more Pokémon based on species name
   *
   * You can provide `take` to limit the amount of Pokémon to return (default: 1), set the offset of where to start with `offset`, and reverse the entire array with `reverse`.
   *
   * You can provide `takeFlavorTexts` to limit the amount of flavour texts to return, set the offset of where to start with `offsetFlavorTexts`, and reverse the entire array with `reverseFlavorTexts`.
   *
   * **Reversal is applied before pagination!**
   */
  getFuzzyPokemon: Array<Pokemon>;
  /** Gets the details on a Pokémon item, using the item name */
  getItem: Item;
  /**
   * Gets the learnsets for a given Pokémon and move.
   *
   * Multiple moves are possible by putting them in an array: `[move1, move2]`.
   *
   * You can also apply a generation filter (only results for the given generation will be returned) with the generation argument
   */
  getLearnset: Learnset;
  /** Gets the details on a Pokémon move, using the move name */
  getMove: Move;
  /**
   * Gets details on a single Pokémon based on species name
   *
   * You can provide `takeFlavorTexts` to limit the amount of flavour texts to return, set the offset of where to start with `offsetFlavorTexts`, and reverse the entire array with `reverseFlavorTexts`.
   *
   * **Reversal is applied before pagination!**
   */
  getPokemon: Pokemon;
  /**
   * Gets details on a single Pokémon based on National Pokédex number
   *
   * You can provide `takeFlavorTexts` to limit the amount of flavour texts to return, set the offset of where to start with `offsetFlavorTexts`, and reverse the entire array with `reverseFlavorTexts`.
   *
   * **Reversal is applied before pagination!**
   */
  getPokemonByDexNumber: Pokemon;
  /** Gets the type matchup data for the given type or types */
  getTypeMatchup: TypeMatchup;
};
// ...
export type Stats = {
  __typename?: 'Stats';
  /** The base attack stat of a Pokémon */
  attack: Scalars['Int']['output'];
  /** The base defense stat of a Pokémon */
  defense: Scalars['Int']['output'];
  /** The base HP stat of a pokémon */
  hp: Scalars['Int']['output'];
  /** The base special attack stat of a Pokémon */
  specialattack: Scalars['Int']['output'];
  /** The base special defense stat of a Pokémon */
  specialdefense: Scalars['Int']['output'];
  /** The base speed stat of a Pokémon */
  speed: Scalars['Int']['output'];
};
// ...
export type AllPokemonQueryQuery = { __typename?: 'Query', getAllPokemon: Array<{ __typename?: 'Pokemon', num: number, key: PokemonEnum, species: string, baseSpecies?: string | null, sprite: string, backSprite: string, height: number, color: string, baseStatsTotal: number, legendary: boolean, mythical: boolean, levellingRate?: string | null, baseStats: { __typename?: 'Stats', attack: number, defense: number, hp: number, specialattack: number, specialdefense: number, speed: number }, catchRate?: { __typename?: 'CatchRate', base: number } | null, types: Array<{ __typename?: 'PokemonType', name: string }>, abilities: { __typename?: 'Abilities', first: { __typename?: 'Ability', key: AbilitiesEnum, shortDesc: string }, hidden?: { __typename?: 'Ability', shortDesc: string, key: AbilitiesEnum, name: string } | null }, flavorTexts: Array<{ __typename?: 'Flavor', flavor: string }> }> };